# Allow subclasses of known types to be encoded with superclass encoder