--- a/changes/1291-StephenBrown2.md
+++ b/changes/1291-StephenBrown2.md
@@ -1,0 +1,1 @@
+Allow subclasses of known types to be encoded with superclass encoder