--- a/docs/examples/exporting_models_json.py
+++ b/docs/examples/exporting_models_json.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timedelta
+import pendulum
 from pydantic import BaseModel
 from pydantic.json import timedelta_isoformat
 
@@ -23,4 +24,20 @@ class WithCustomEncoders(BaseModel):
         }
 
 m = WithCustomEncoders(dt=datetime(2032, 6, 1), diff=timedelta(hours=100))
+print(m.json())
+# (returns a str)
+class WithSubclasses(BaseModel):
+    pdt: pendulum.DateTime
+    pdiff: pendulum.Duration
+
+    class Config:
+        json_encoders = {
+            datetime: lambda v: v.timestamp(),
+            timedelta: timedelta_isoformat,
+        }
+
+m = WithSubclasses(
+    pdt=pendulum.datetime(2032, 6, 1),
+    pdiff=pendulum.duration(months=3, days=2, hours=1),
+)
 print(m.json())