--- a/docs/usage/exporting_models.md
+++ b/docs/usage/exporting_models.md
@@ -60,10 +60,7 @@ _(This script is complete, it should run "as is")_
 
 The `.json()` method will serialise a model to JSON. Typically, `.json()` in turn calls `.dict()` and
 serialises its result. (For models with a [custom root type](models.md#custom-root-types), after calling `.dict()`,
-only the value for the `__root__` key is serialised.)
-
-Serialisation can be customised on a model using the `json_encoders` config property; the keys should be types, and
-the values should be functions which serialise that type (see the example below).
+only the value for the `__root__` key is serialised)
 
 Arguments:
 
@@ -81,18 +78,63 @@ Arguments:
   encoder designed to take care of all common types
 * `**dumps_kwargs`: any other keyword arguments are passed to `json.dumps()`, e.g. `indent`.
 
-Example:
+*pydantic* can serialise many commonly used types to JSON (e.g. `datetime`, `date` or `UUID`) which would normally
+fail with a simple `json.dumps(foobar)`.
 
 ```py
 {!.tmp_examples/exporting_models_json.py!}
 ```
 _(This script is complete, it should run "as is")_
 
+### `json_encoders`
+
+Serialisation can be customised on a model using the `json_encoders` config property; the keys should be types, and
+the values should be functions which serialise that type (see the example below):
+
+```py
+{!.tmp_examples/exporting_models_json_encoders.py!}
+```
+_(This script is complete, it should run "as is")_
+
 By default, `timedelta` is encoded as a simple float of total seconds. The `timedelta_isoformat` is provided
 as an optional alternative which implements ISO 8601 time diff encoding.
 
-See [below](#custom-json-deserialisation) for details on how to use other libraries for more performant JSON encoding
-and decoding.
+### Serialising subclasses
+
+!!! note
+    New in version **v1.5**.
+    
+    Subclasses of common types were not automatically serialised to JSON before **v1.5**.
+
+Subclasses of common types are automatically encoded like their super-classes:
+
+```py
+{!.tmp_examples/exporting_models_json_subclass.py!}
+```
+_(This script is complete, it should run "as is")_
+
+### Custom JSON (de)serialisation
+
+To improve the performance of encoding and decoding JSON, alternative JSON implementations
+(e.g. [ujson](https://pypi.python.org/pypi/ujson)) can be used via the
+`json_loads` and `json_dumps` properties of `Config`.
+
+```py
+{!.tmp_examples/exporting_models_ujson.py!}
+```
+_(This script is complete, it should run "as is")_
+
+`ujson` generally cannot be used to dump JSON since it doesn't support encoding of objects like datetimes and does
+not accept a `default` fallback function argument. To do this, you may use another library like
+[orjson](https://github.com/ijl/orjson).
+
+```py
+{!.tmp_examples/exporting_models_orjson.py!}
+```
+_(This script is complete, it should run "as is")_
+
+Note that `orjson` takes care of `datetime` encoding natively, making it faster than `json.dumps` but
+meaning you cannot always customise the encoding using `Config.json_encoders`.
 
 ## `pickle.dumps(model)`
 
@@ -124,26 +166,3 @@ member of a list or tuple, the dictionary key `'__all__'` can be used as follows
 ```
 
 The same holds for the `json` and `copy` methods.
-
-## Custom JSON (de)serialisation
-
-To improve the performance of encoding and decoding JSON, alternative JSON implementations
-(e.g. [ujson](https://pypi.python.org/pypi/ujson)) can be used via the
-`json_loads` and `json_dumps` properties of `Config`.
-
-```py
-{!.tmp_examples/exporting_models_ujson.py!}
-```
-_(This script is complete, it should run "as is")_
-
-`ujson` generally cannot be used to dump JSON since it doesn't support encoding of objects like datetimes and does
-not accept a `default` fallback function argument. To do this, you may use another library like
-[orjson](https://github.com/ijl/orjson).
-
-```py
-{!.tmp_examples/exporting_models_orjson.py!}
-```
-_(This script is complete, it should run "as is")_
-
-Note that `orjson` takes care of `datetime` encoding natively, making it faster than `json.dumps` but
-meaning you cannot always customise the encoding using `Config.json_encoders`.