--- a/pydantic/json.py
+++ b/pydantic/json.py
@@ -55,9 +55,11 @@ def pydantic_encoder(obj: Any) -> Any:
 
     # Check the class type and its superclasses for a matching encoder
     for base in obj.__class__.__mro__[:-1]:
-        encoder = ENCODERS_BY_TYPE.get(base)
-        if encoder:
-            return encoder(obj)
+        try:
+            encoder = ENCODERS_BY_TYPE[base]
+        except KeyError:
+            continue
+        return encoder(obj)
     else:  # We have exited the for loop without finding a suitable encoder
         raise TypeError(f"Object of type '{obj.__class__.__name__}' is not JSON serializable")
 
@@ -65,9 +67,11 @@ def pydantic_encoder(obj: Any) -> Any:
 def custom_pydantic_encoder(type_encoders: Dict[Any, Callable[[Type[Any]], Any]], obj: Any) -> Any:
     # Check the class type and its superclasses for a matching encoder
     for base in obj.__class__.__mro__[:-1]:
-        encoder = type_encoders.get(base)
-        if encoder:
-            return encoder(obj)
+        try:
+            encoder = type_encoders[base]
+        except KeyError:
+            continue
+        return encoder(obj)
     else:  # We have exited the for loop without finding a suitable encoder
         return pydantic_encoder(obj)