## Change Summary

After further research, it seems my aside on `__mro__` usage in #1281 was unfounded and unnecessary, and the only other place I figured this change might be useful ([validators.py](https://github.com/samuelcolvin/pydantic/blob/8727914ced9d4546f0f95a453aff78f03b778275/pydantic/validators.py#L557-L568)) seems to already be handled alright. Thus, I'm submitting this PR as-is to resolve the issue found with JSON serialization of subclasses of known types.

While adding documentation, I found that the custom `json_encoders` were being ignored, so I added the same fix to `custom_pydantic_encoder` as well.

## Related issue number

Resolves #1281 

## Checklist

* [x] Unit tests for the changes exist
* [x] Tests pass on CI and coverage remains at 100%
* [x] Documentation reflects the changes where applicable
* [x] `changes/<pull request or issue id>-<github username>.md` file added describing change
  (see [changes/README.md](https://github.com/samuelcolvin/pydantic/blob/master/changes/README.md) for details)