@samuelcolvin 2020-03-06
I'm concerned about the performance impact of this as discussed on #1157, I hope to provide a better solution for this kind of thing in future.

You can achieve what you want by using your own `config.json_encoders`