@StephenBrown2 2020-03-06
> I'm concerned about the performance impact of this as discussed on #1157, I hope to provide a better solution for this kind of thing in future.

I'm not sure what performance impacts you are seeing, I am seeing the same or slightly faster performance with my tests (cleaning and compiling before each run):
```
Benchmark on master:
python benchmarks/run.py pydantic-only
testing pydantic, 10 times each
    pydantic ( 1/10) time=0.437s, success=50.10%
    pydantic ( 2/10) time=0.451s, success=50.10%
    pydantic ( 3/10) time=0.437s, success=50.10%
    pydantic ( 4/10) time=0.428s, success=50.10%
    pydantic ( 5/10) time=0.433s, success=50.10%
    pydantic ( 6/10) time=0.432s, success=50.10%
    pydantic ( 7/10) time=0.431s, success=50.10%
    pydantic ( 8/10) time=0.419s, success=50.10%
    pydantic ( 9/10) time=0.531s, success=50.10%
    pydantic (10/10) time=0.449s, success=50.10%
    pydantic best=0.419s, avg=0.445s, stdev=0.032s

    pydantic best=69.874μs/iter avg=74.140μs/iter stdev=5.273μs/iter version=1.4a1

Benchmark on encode_known_subclasses:
python benchmarks/run.py pydantic-only
testing pydantic, 10 times each
    pydantic ( 1/10) time=0.467s, success=50.10%
    pydantic ( 2/10) time=0.471s, success=50.10%
    pydantic ( 3/10) time=0.431s, success=50.10%
    pydantic ( 4/10) time=0.427s, success=50.10%
    pydantic ( 5/10) time=0.434s, success=50.10%
    pydantic ( 6/10) time=0.434s, success=50.10%
    pydantic ( 7/10) time=0.423s, success=50.10%
    pydantic ( 8/10) time=0.417s, success=50.10%
    pydantic ( 9/10) time=0.433s, success=50.10%
    pydantic (10/10) time=0.454s, success=50.10%
    pydantic best=0.417s, avg=0.439s, stdev=0.019s

    pydantic best=69.490μs/iter avg=73.185μs/iter stdev=3.087μs/iter version=1.4a1
```

Using this script to make sure things were fair
<details><summary><code>compare_branch_perf.sh</code>: </summary>

```bash
#!/usr/bin/env bash
branch=${1:-$(git branch --show-current)}
export BENCHMARK_REPEATS=${2:-10}
output=$(mktemp -t "benchmark-${branch}")

if [[ "${branch}" == "master" ]]; then
  echo "Currently on master, and no branch specified."
  echo "Either check out the branch to compare or specify it as the first argument."
  break
fi

echo "Preparing to compare benchmarks on master vs ${branch}, ${BENCHMARK_REPEATS} repetitions"

git checkout master
echo "Cleaning and building with cython"
make clean build-cython 2>&1 | awk '{printf "."; system("")} END{print "\n"}'
echo "Benchmarking master..."
echo "Benchmark on master:" > ${output}
make benchmark-pydantic | tee -a ${output}

git checkout ${branch}
echo "Cleaning and building with cython"
make clean build-cython 2>&1 | awk '{printf "."; system("")} END{print "\n"}'
echo "Benchmarking ${branch}..."
echo -e "\nBenchmark on ${branch}:" >> ${output}
make benchmark-pydantic | tee -a ${output}

clear
cat ${output}
```

</details>

Can you show performance benchmarks that show significant degredation?
[This branch in Travis](https://travis-ci.org/samuelcolvin/pydantic/jobs/658387558#L656): `pydantic best=56.288μs/iter avg=56.288μs/iter stdev=0.000μs/iter version=1.4a1`
[Master in Travis](https://travis-ci.org/samuelcolvin/pydantic/jobs/659158749#L670): `                                pydantic best=56.039μs/iter avg=56.039μs/iter stdev=0.000μs/iter version=1.4a1`

A difference of `0.249μs`.

> You can achieve what you want by using your own `config.json_encoders`

Yes, but that's exactly the point of this. In case I want to use `pendulum` objects instead of naive datetimes everywhere, I would have to add the custom `Config.json_encoders` to every single model that has one. This is a simple, performant solution that allows me to use subclasses without having to add a lot of boilerplate to my code, and also resolves #1157.

Of note, I attempted to try a faster solution that still compared all the elements but shortcircuited using set logic to avoid a loop, and came out just a tiny bit slower (a rounding error), using this logic:
```py
    if set(obj.__class__.__mro__) & set(ENCODERS_BY_TYPE.keys()):
        return next(ENCODERS_BY_TYPE[base](obj) for base in obj.__class__.__mro__ if base in ENCODERS_BY_TYPE)
    else:
        raise TypeError(f"Object of type '{obj.__class__.__name__}' is not JSON serializable")
```

```
Benchmark on master:
python benchmarks/run.py pydantic-only
testing pydantic, 10 times each
    pydantic ( 1/10) time=0.444s, success=50.10%
    pydantic ( 2/10) time=0.422s, success=50.10%
    pydantic ( 3/10) time=0.415s, success=50.10%
    pydantic ( 4/10) time=0.415s, success=50.10%
    pydantic ( 5/10) time=0.413s, success=50.10%
    pydantic ( 6/10) time=0.420s, success=50.10%
    pydantic ( 7/10) time=0.455s, success=50.10%
    pydantic ( 8/10) time=0.432s, success=50.10%
    pydantic ( 9/10) time=0.420s, success=50.10%
    pydantic (10/10) time=0.442s, success=50.10%
    pydantic best=0.413s, avg=0.428s, stdev=0.015s

    pydantic best=68.833μs/iter avg=71.283μs/iter stdev=2.440μs/iter version=1.4a1

Benchmark on encode_known_subclasses:
python benchmarks/run.py pydantic-only
testing pydantic, 10 times each
    pydantic ( 1/10) time=0.464s, success=50.10%
    pydantic ( 2/10) time=0.425s, success=50.10%
    pydantic ( 3/10) time=0.404s, success=50.10%
    pydantic ( 4/10) time=0.449s, success=50.10%
    pydantic ( 5/10) time=0.429s, success=50.10%
    pydantic ( 6/10) time=0.438s, success=50.10%
    pydantic ( 7/10) time=0.446s, success=50.10%
    pydantic ( 8/10) time=0.421s, success=50.10%
    pydantic ( 9/10) time=0.416s, success=50.10%
    pydantic (10/10) time=0.404s, success=50.10%
    pydantic best=0.404s, avg=0.430s, stdev=0.020s

    pydantic best=67.315μs/iter avg=71.630μs/iter stdev=3.276μs/iter version=1.4a1
```