@StephenBrown2 2020-03-06
As it is, FastAPI implements [nearly the same](https://github.com/samuelcolvin/pydantic/issues/1157#issuecomment-572511480) logic [in its `jsonable_encoder`](https://github.com/tiangolo/fastapi/blob/025b38df40a18f94322fab445aee13024fe783da/fastapi/encoders.py#L128-L140), and while I would rather not add a custom encoder to all my classes, I suppose I could subclass `BaseModel` and add it there.

I would argue it is not "always going to be slower", as it will be exactly the same performance for native python types (as a loop over a single element array `O(n/n)` is the same as accessing a value by key `O(1)`), and only get slower as the inheritance chain grows. Thus on average, the performance hit would be negligible as demonstrated, and only applicable if one has a very deeply inherited object. In which case, one should certainly create a custom encoder, which will speed up the execution back to native levels.

Thank you for your time and consideration, and the efforts you've put into making this one of the best libraries for what it does.